changes to crypt package
